Combine remote display with background video.
Video (background.mp4, not included) autoplays with the PicoPew small in the corner. When keys are pressed on the PicoPew, it expands to fill the screen, 1 minute after the last key press it recedes to the corner again.

Video is paused while the PicoPew is large because my Raspberry Pi 3B+ struggles with rendering both video and PicoPew at a decent frame rate.